refactor gateway to rpc
